Added `de/sets` module
